Remove article info block from templates
This site is not a blog, so articles will be revised as often as I find
the time and inclination to do so. Therefore, article publication dates
are not useful.

This site is not collaboratively edited. Therefore, author credits are
in the footer, and displaying them elsewhere on the page is not useful.

That leaves only the translations block, and I am monolingual.