Explain what to do when a program surprises you
It's a first pass, but I think it gets the basics right, so it's going
live.

Strongly inspired by me making a mistake in how I handled a weird issue
that's no longer manifesting, then taking the time to articulate clearly
how situations like that should be handled in general.