Explain why I announce goals for the day (tersely)
I wrote a better explanation in a previous commit message, so perhaps I
should mine that someday.

...man, now I'm thinking back to the days when I believed huge commit
messages were a good thing, before I realized they usually just contain
information that belongs in the project itself (details on the
workaround, the business justification for a change, etc).

Ah, the naivety of youth.