Add more constraints to design for detecting nicknames
I feel it shows nicely how impossible a seemingly-simple problem can be,
and also illustrates how pointless nicknames in a system actually are.
You're doing all this so the program can prefix error messages with your
name, in a futile attempt to make it feel less like a robot being a
robot?