Add POCs to my task achievement routine
Today I worked on code that proved irrelevant as I had not POC'd the
core parts of the feature I was working on.

If I always started by POC'ing the essential components, such surprises
would happen far less often, and I would save a lot of time in the long
term.