Clarify the nature of people taking precedence over goals
Long-term, I suspect companies do need to prioritize big-picture goals
over individuals, even if one of those goals is actually giving
employees the chance to self-actualize. Otherwise there's a heightened
risk of existential failure due to insufficient focus.

It's in the day-to-day that you should prioritize people over goals, in
part because the big-picture goals are usually best attained by doing
that.