Note that I should link qntm.org's argument against abolishing timezones
The author makes the case well and I should indicate my support for the
position.

As the comments show, while there are conceivable workarounds for the
problems the authors points out in a universal timezone, those
workarounds boil down to "reinvent timezones informally so people can
think reasonably about other people's lived experience of time".

That sounds like a stupid plan to me.