Note that dependency declarations should be organized per-feature
While defining an Ansible task for a server role at $DAYJOB, I realized
that the *why* of any third-partyh dependency is necessarily a
user-facing feature or developer tooling.

If it's neither of those, it shouldn't exist.

Therefore, we should structure our dependency management in such a way
as to capture this valuable information that is usually lost to the
mists of time.