More changes to the daily routine
I'm allowing todo lists to exist in the daily log again, as an
experiment in being faster when I need to capture "thing to follow up
on" without losing my flow.

Not sure it's a good idea - we'll see how it works out.

This change also moves a bunch of stuff out from the triage section that
didn't belong there.