Add some notes on how to handle errors
These are inspired by a recent failure of mine, in which I spent hours
trying to figure out what was causing an apparently-silent error.

It turned out I was explicitly logging the error condition at debug
level and not re-throwing the exception after the log statement
executed. I missed the actual error message in the midst of all the
other debugging noise.

I believe if I follow the rules I've defined here, I won't do that to
myself (or another programmer) again.